+++ NEWS.rst
@@ -1,3 +1,10 @@
+Changes in 3.7.7 (released 03/01/2012)
+--------------------------------------
+
+*	Added the attributes ``allowFullScreen`` and ``flashvars`` to
+	:class:`ll.xist.ns.html.embed`.
+
+
 Changes in 3.7.6 (released 05/14/2010)
 --------------------------------------